+++ packages/liveblocks-react-comments/src/primitives/Composer/types.ts
@@ -1,5 +1,10 @@
 import type { CommentBody } from "@liveblocks/core";
-import type { ComponentPropsWithoutRef, ComponentType, FormEvent } from "react";
+import type {
+  ComponentPropsWithoutRef,
+  ComponentType,
+  FormEvent,
+  ReactNode,
+} from "react";
 import type {
   RenderElementProps,
   RenderElementSpecificProps,
@@ -38,7 +43,7 @@ export interface ComposerRenderLinkProps {
    *
    * @example "www.example.com", "a link", etc.
    */
-  children: string;
+  children: ReactNode;
 }
 
 export type ComposerMentionProps = ComponentPropsWithSlot<"span">;